git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@989 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/datamanagement/stringmemory.cpp
@@ -376,11 +376,11 @@ string StringMemory::maxString(unsigned int i1, unsigned int i2, unsigned int nC
 string StringMemory::maxString(VectorIndex _vLine, VectorIndex _vCol)
 {
     // Ensure that the selected column exists
-    if (_vCol.front() >= _stringIntMem.sStrings.size())
+    if (_vCol.front() >= (int)_stringIntMem.sStrings.size())
         return "";
 
     // Fill the second line index automatically
-    if (_vLine.isOpenEnd() || _vLine.last() > _stringIntMem.sStrings[_vCol.front()].size())
+    if (_vLine.isOpenEnd() || _vLine.last() > (int)_stringIntMem.sStrings[_vCol.front()].size())
         _vLine.setRange(0, _stringIntMem.sStrings[_vCol.front()].size()-1);
 
     // Return an empty string, if the second column does not exist
@@ -432,11 +432,11 @@ string StringMemory::minString(unsigned int i1, unsigned int i2, unsigned int nC
 string StringMemory::minString(VectorIndex _vLine, VectorIndex _vCol)
 {
     // Ensure that the selected column exists
-    if (_vCol.front() >= _stringIntMem.sStrings.size())
+    if (_vCol.front() >= (int)_stringIntMem.sStrings.size())
         return "";
 
     // Fill the second line index automatically
-    if (_vLine.isOpenEnd() || _vLine.last() > _stringIntMem.sStrings[_vCol.front()].size())
+    if (_vLine.isOpenEnd() || _vLine.last() > (int)_stringIntMem.sStrings[_vCol.front()].size())
         _vLine.setRange(0, _stringIntMem.sStrings[_vCol.front()].size()-1);
 
     // Return an empty string, if the second column does not exist
@@ -487,11 +487,11 @@ string StringMemory::sumString(unsigned int i1, unsigned int i2, unsigned int nC
 string StringMemory::sumString(VectorIndex _vLine, VectorIndex _vCol)
 {
     // Ensure that the selected column exists
-    if (_vCol.front() >= _stringIntMem.sStrings.size())
+    if (_vCol.front() >= (int)_stringIntMem.sStrings.size())
         return "";
 
     // Fill the second line index automatically
-    if (_vLine.isOpenEnd() || _vLine.last() > _stringIntMem.sStrings[_vCol.front()].size())
+    if (_vLine.isOpenEnd() || _vLine.last() > (int)_stringIntMem.sStrings[_vCol.front()].size())
         _vLine.setRange(0, _stringIntMem.sStrings[_vCol.front()].size()-1);
 
     // Return an empty string, if the second column does not exist
